Refractoring of sidenav into "nav," stateless components
