--- pages/components/customTable/styles.less
@@ -1,4 +1,0 @@
-.container {
-  padding: 15px;
-  height: 100vh;
-}
