+++ pages/components/nav/header/styles.less
@@ -1,0 +1,22 @@
+@import '../../../global-styles';
+.container {
+  position: absolute;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  z-index: 1000;
+  .navMenuContainer {
+    min-width: @sidenav-width;
+    padding: 30px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    .logo {
+      transition: fill 0.25s;
+    }
+    .logoDark {
+      .logo();
+      fill: @primaryDark;
+    }
+  }
+}
