--- pages/components/sidenav/navItem/styles.less
@@ -1,91 +1,0 @@
-@import '../../../global-styles';
-
-.navItem {
-  @button-base();
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-
-  padding-left: 30px;
-  width: 100%;
-  height: 75px;
-  font-weight: 300;
-
-  &:hover {
-      background-color: @primaryAccentDark;
-
-      .content:after {
-          outline-style: solid;
-          outline-color: @primary;
-          outline-width: thin;
-          width: 100%;
-      }
-
-      .dropdownIconWrapper {
-          .dropdownIcon {
-              fill: @primary;
-          }
-          .dropdownIconActive {
-              fill: @primary;
-          }
-      }
-  }
-
-  .content {
-      font-size: 16px;
-      color: @primaryLight;
-      display: inline-block;
-      margin-top: 5px;
-
-      &:after {
-          content: "";
-          position: relative;
-          display: block;
-          margin-top: 5px;
-          height: 0;
-          width: 0;
-          transition: width .25s;
-      }
-  }
-
-  .dropdownIconWrapper {
-      flex: 1;
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-end;
-      align-self: center;
-  
-      .dropdownIcon {
-          fill: @primaryLight;
-          margin-right: 30px;
-          overflow: visible;
-          transition: 0.25s;
-      }
-
-      .dropdownIconActive {
-          .dropdownIcon();
-          transform: rotate(180deg);
-      }
-  }
-}
-
-.navSubItem {
-  .navItem();
-  align-items: center;
-  height: 60px;
-  padding-left: 60px;
-
-  .content {
-      font-size: 14px;
-  }
-
-  &:before {
-      content: "";
-      position: absolute;
-      top: 0;
-      left: 60px;
-      width: 100%;
-      border-top: solid 1px @primaryAccentDark;
-  }
-}
