+++ pages/global-styles.less
@@ -1,21 +1,19 @@
 @primaryDark: #333;
 @primaryAccentDark: #444;
 @secondaryDark: #666;
-
 @primaryLight: #fbfbfb;
-
 @primary: #F2D600;
-@background:#f7f7f7;
-
+@background: #f7f7f7;
 //for ant design theme override
 // @primary-color: @primary;
-
 @button-base: {
-    cursor: pointer;
-    background-color: transparent;
-    outline: none;
-    border: none;
-    overflow: hidden;
-    padding: 0;
-    margin: 0;
+  cursor: pointer;
+  background-color: transparent;
+  outline: none;
+  border: none;
+  overflow: hidden;
+  padding: 0;
+  margin: 0;
 };
+
+@sidenav-width: 300px;
