+++ pages/layout/styles.less
@@ -6,5 +6,6 @@
     flex: 1;
     overflow-y: scroll;
     height: 100vh;
+    padding-top: 90px;
   }
 }
